==> qbs_scripts/01 run.R <==
# 01 RUN ALL ####
## WHO: JOE COLLINS
## WHAT: Run QBS
## LAST EDIT: 25/01/2024
####

getwd()


## 02 PACKAGES ####

source(file = "qbs_scripts/02 packages.R")

## 03 DATA ####

source(file = "qbs_scripts/03 qbs.data.R")

## 04 DISTRIBUTION AND NORMALITY ####

source(file = "qbs_scripts/Scripts/04 distribution.R")

## 05 STATISTICS ####

source(file = "qbs_scripts/Scripts/05 statistics.R")

## 06 PCA ####

source(file = "qbs_scripts/Scripts/06 PCA.R")
